Refactor console command handling to remove deprecated binary bank support
This commit removes the legacy binary bank frame handling from the console command processing in `channel_console.c`, transitioning to a purely ASCII-based command structure. Key changes include:

- Eliminated the `Console_ApplyBinBank` function and related binary mode logic.
- Updated the command buffer size to accommodate new ASCII command formats.
- Revised documentation to reflect the removal of binary bank frames and the emphasis on fractional Hz input.

These changes aim to simplify the command interface and enhance compatibility with the current audio processing architecture.

+++ libs/rs485/include/rs485/session.hpp
@@ -65,9 +65,10 @@ public:
 
   ExchangeResult Exec(Target target, const std::string &command);
 
-  /** Integer Hz; 0 = off. Slot 0..15. */
-  ExchangeResult SetNote(uint8_t slot, uint16_t hz);
-  ExchangeResult NoteOff(uint8_t slot) { return SetNote(slot, 0); }
+  /** Frequency in Hz (fractional OK); 0 = off. Slot 0..15.
+   *  Integer-only Hz breaks ET octaves (e.g. C4→262 / C5→523 → 1 Hz beat). */
+  ExchangeResult SetNote(uint8_t slot, double hz);
+  ExchangeResult NoteOff(uint8_t slot) { return SetNote(slot, 0.0); }
   ExchangeResult Silence();
   ExchangeResult Gain(uint8_t ch, uint8_t atten_db);
